pyproject.toml
[tool.poetry]
name = "nz-covid-scanner"
version = "0.1.0"
description = ""
authors = ["Sean McGrath <sean@mcgrath.net.nz>"]

[tool.poetry.dependencies]
python = "^3.9"
cryptography = "3.3.2"
cwt = "^1.4.2"
requests = "^2.26.0"
Pillow = "^8.4.0"
pyserial = "^3.5"
pygame = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.12b0"

[tool.poetry.scripts]
nz_covid_scanner = "nz_covid_scanner.main:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
